python_service/requirements.txt
moviepy
edge-tts
openai-whisper 
pillow 
requests 
fastapi 
uvicorn